requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
bcrypt==4.0.1
cryptography==41.0.7
python-dotenv==1.0.0
Werkzeug==2.3.7
PyMySQL==1.1.0
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-flask==1.2.0
coverage==7.3.2
black==23.11.0
flake8==6.1.0
bleach==6.1.0
boto3==1.34.0
mypy==1.7.1